requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
attrs==25.4.0
bracex==2.6
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dockerfile-parse==2.0.1
docstring_parser==0.17.0
e2b==2.8.1
e2b-code-interpreter==2.4.1
fonttools==4.61.1
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-anthropic==1.3.0
langchain-core==1.2.4
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.4.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pandoc==2.4
pillow==12.0.0
plumbum==1.10.0
ply==3.11
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
wcmatch==10.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
chainlit=1.12.3
langchain=0.0.208
anthropic-sdk=0.3.0
e2b-code-interpreter=2.4.1
google-genai=1.56.0
pandas-stubs=1.5.3.230531
matplotlib-stubs=0.1.3
python-dotenv=1.2.1
rich=14.2.0
streamlit=1.30.1
# Export documents
python-docx==1.1.0
weasyprint==60.2
openpyxl==3.1.2

# Pour le déploiement
watchdog==3.0.0
logging==0.5.1
gunicorn==20.1.0
threating==0.1.3
boto3==1.30.25
typing==4.15.0
botocore==1.38.25
google.generativeai==1.56.0